+++ core/src/main/java/arez/Spy.java
@@ -122,16 +122,6 @@ public interface Spy
   Collection<ComputedValueInfo> findAllTopLevelComputedValues();
 
   /**
-   * Return true if the specified ObservableValue has a mutator.
-   * This method should not be invoked if {@link Arez#arePropertyIntrospectorsEnabled()} returns false.
-   *
-   * @param <T>             The type of the value that is observableValue.
-   * @param observableValue the ObservableValue.
-   * @return true if a mutator is available.
-   */
-  <T> boolean hasMutator( @Nonnull ObservableValue<T> observableValue );
-
-  /**
    * Set the value of the specified ObservableValue.
    * This method should only be invoked if {@link Arez#arePropertyIntrospectorsEnabled()} returns true
    * and {@link #hasMutator(ObservableValue)} for the same element returns true.